content/media/video/test/test_wav_timeupdate1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Wave Media test: timeupdate during playback</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test if timeupdate is dispatched and the current time during the
// update is increasing.
var lastTime = 0.0;
var completed = false;

function startTest() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  v.play();
  return false;
}

function timeUpdated() {
  if (completed)
    return false;

  var v = document.getElementById('v');

  // Check that duration metadata is correct during the first timeupdate.
  if (lastTime == 0) {
    ok(v.duration > 0.9 && v.duration < 1.1, "Duration should be around 1.0: " + v.duration);
  }

  ok(v.currentTime > lastTime,
     "Check currentTime of " +
     v.currentTime +
     " is greater than last time of " +
     lastTime);
  lastTime = v.currentTime;
  return false;
}

function playbackEnded() {
  if (completed)
    return false

  completed = true;
  SimpleTest.finish();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<audio id='v'
       src='r11025_u8_c1.wav'
       onloadedmetadata='return startTest();'
       onended='return playbackEnded();'
       ontimeupdate='return timeUpdated();'></audio>
</body>
</html>
